// plugin/songloft/.research/miot-src/types.ts
export type PlayState = 'idle' | 'playing' | 'paused' | 'stopped';

/** 播放模式枚举 */
export type PlayMode = 'order' | 'random' | 'single' | 'loop';